requirements.txt
transformers>=4.41.0
torch>=2.2.0
Pillow>=10.0.0